refactor(backend): 迁移 ID 类型 Integer→Long (ASSIGN_ID) + Jackson Long→String
数据库主键已改为雪花 bigint，后端联动：
- 9 个实体 ID 字段 Integer→Long，@TableId 由 IdType.AUTO 改 ASSIGN_ID
  （雪花 ID 由 MyBatis-Plus 应用层生成，库表无自增）
- mapper/DTO/VO/控制器/service/business 全部 ID 签名与字段→Long；
  删除死代码 PlaylistsMapper.insertAndGetId
- QueuesBusinessImpl：Map 强转 (Integer)→(Long)（queue_id/song_id），
  removedPos/finalPosition 等位置字段保持 int
- JWT：createToken(Long)，过滤器 user_id 读取改 longValue()；request attribute 变 Long
- SongScannerTask/CodeGenerator 联动；PaginationVO 三个 long 字段改 int
- 新增 JacksonConfig：Long/long 全局序列化为字符串，避免雪花 ID 超过
  JS 2^53 精度（JSON 里 ID 为字符串）

+++ backend/src/main/java/com/kyf/mp/javaserver/common/auth/JwtAuthenticationFilter.java
@@ -47,7 +47,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
             if (!(rawUserId instanceof Number userId)) {
                 return;
             }
-            Integer id = userId.intValue();
+            Long id = userId.longValue();
             UsernamePasswordAuthenticationToken authentication =
                     UsernamePasswordAuthenticationToken.authenticated(id, null, List.of());
             SecurityContextHolder.getContext().setAuthentication(authentication);
