+++ backend/src/main/java/com/kyf/mp/javaserver/modules/playlistmodule/controller/PlaylistsController.java
@@ -44,7 +44,7 @@ public class PlaylistsController {
             @RequestParam("cover_image") @NotNull(message = "请上传歌单封面") MultipartFile file,
             @RequestParam("name") @NotBlank(message = "歌单名称不能为空") String name,
             @RequestParam(value = "description", required = false) String description,
-            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Integer userId) {
+            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Long userId) {
         return ResultModel.success(playlistsService.createPlaylist(file, name, description, userId));
     }
 
@@ -52,18 +52,18 @@ public class PlaylistsController {
     @PatchMapping
     public ResultModel<PlaylistActionVO> editPlaylist(
             @RequestParam(value = "cover_image", required = false) MultipartFile file,
-            @RequestParam("playlist_id") @NotNull(message = "歌单ID不能为空") Integer playlistId,
+            @RequestParam("playlist_id") @NotNull(message = "歌单ID不能为空") Long playlistId,
             @RequestParam(value = "name", required = false) @NotBlank(message = "歌单名称不能为空") String name,
             @RequestParam(value = "description", required = false) String description,
-            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Integer userId) {
+            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Long userId) {
         return ResultModel.success(playlistsService.editPlaylist(file, playlistId, name, description, userId));
     }
 
     // 3. 删除歌单
     @DeleteMapping("/{id}")
     public ResultModel<Void> deletePlaylist(
-            @PathVariable("id") @NotNull(message = "歌单ID不能为空") Integer id,
-            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Integer userId) {
+            @PathVariable("id") @NotNull(message = "歌单ID不能为空") Long id,
+            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Long userId) {
         playlistsService.deletePlaylist(id, userId);
         return ResultModel.success(null);
     }
@@ -71,31 +71,31 @@ public class PlaylistsController {
     // 4. 获取我的歌单列表
     @GetMapping
     public ResultModel<MyPlaylistsVO> getMyPlaylists(
-            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Integer userId) {
+            @RequestAttribute(value = "userId", required = false) @NotNull(message = "用户ID不能为空") Long userId) {
         return ResultModel.success(playlistsService.getMyPlaylists(userId));
     }
 
     // 5. 获取歌单详情（含歌曲）
     @GetMapping("/{id}")
     public ResultModel<PlaylistDetailVO> getPlaylistById(
-            @PathVariable("id") @NotNull(message = "歌单ID不能为空") Integer id,
-            @RequestAttribute("userId") Integer userId) {
+            @PathVariable("id") @NotNull(message = "歌单ID不能为空") Long id,
+            @RequestAttribute("userId") Long userId) {
         return ResultModel.success(playlistsService.getPlaylistDetail(id, userId));
     }
 
     // 6. 点赞/取消点赞
     @PostMapping("/{id}/likes")
     public ResultModel<Void> like(
-            @PathVariable("id") @NotNull(message = "参数不完整") Integer id,
-            @RequestAttribute(value = "userId", required = false) @NotNull(message = "参数不完整") Integer userId) {
+            @PathVariable("id") @NotNull(message = "参数不完整") Long id,
+            @RequestAttribute(value = "userId", required = false) @NotNull(message = "参数不完整") Long userId) {
         playlistsService.toggleLike(id, userId, true);
         return ResultModel.success(null);
     }
 
     @DeleteMapping("/{id}/unlikes")
     public ResultModel<Void> unlike(
-            @PathVariable("id") @NotNull(message = "参数不完整") Integer id,
-            @RequestAttribute(value = "userId", required = false) @NotNull(message = "参数不完整") Integer userId) {
+            @PathVariable("id") @NotNull(message = "参数不完整") Long id,
+            @RequestAttribute(value = "userId", required = false) @NotNull(message = "参数不完整") Long userId) {
         playlistsService.toggleLike(id, userId, false);
         return ResultModel.success(null);
     }
@@ -103,15 +103,15 @@ public class PlaylistsController {
     // 7. 歌单添加/移除歌曲
     @PostMapping("/{playlist_id}/songs/{song_id}")
     public ResultModel<AddSongToPlaylistVO> addSong(
-            @PathVariable("playlist_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Integer playlistId,
-            @PathVariable("song_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Integer songId) {
+            @PathVariable("playlist_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Long playlistId,
+            @PathVariable("song_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Long songId) {
         return ResultModel.success(playlistsService.addSongToPlaylist(playlistId, songId));
     }
 
     @DeleteMapping("/{playlist_id}/songs/{song_id}")
     public ResultModel<Void> removeSong(
-            @PathVariable("playlist_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Integer playlistId,
-            @PathVariable("song_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Integer songId) {
+            @PathVariable("playlist_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Long playlistId,
+            @PathVariable("song_id") @NotNull(message = "歌单ID或歌曲ID不能为空") Long songId) {
         playlistsService.removeSongFromPlaylist(playlistId, songId);
         return ResultModel.success(null);
     }
